requirements.txt
langchain_chroma==1.1.0
langchain_community==0.4.1
langchain_core==1.3.0
pydantic==2.13.2
python-dotenv==1.2.2
streamlit==1.56.0

langchain
langsmith
chromadb
numpy
pandas
wordcloud
matplotlib
